pamplets/requirements.txt
mistralai == 1.6.0
python-dotenv == 1.1.0
fastapi == 0.115.12
uvicorn == 0.34.0
python-multipart==0.0.20
pymongo == 4.11.3



